s3: split out config types into separate header
This is to reduce other modules pulling in whole
trees of http client headers because they needed
a simple thing like s3::bucket name in their config
structure.
